Enqueue heartbeats in priority queue for checking
Enqueues, or re-enqueues heartbeats from nodes as they arrive into a
priority queue.

+++ pkg/lib/collections/hashed_priority_queue.go
@@ -33,7 +33,7 @@ func (q *HashedPriorityQueue[K, T]) Contains(id K) bool {
 
 // Enqueue will add the item specified by `data` to the queue with the
 // the priority given by `priority`.
-func (q *HashedPriorityQueue[K, T]) Enqueue(data T, priority int) {
+func (q *HashedPriorityQueue[K, T]) Enqueue(data T, priority int64) {
 	q.mu.Lock()
 	defer q.mu.Unlock()
